# vehicle-app/api.yaml
openapi: 3.0.0
info:
  title: Vehicle Parking App API
  version: 2.0.0
  description: >
    REST API surface for the Vehicle Parking App. All endpoints return JSON.
    Authentication is handled via Flask sessions, so requests from the SPA must
    send cookies (`withCredentials: true` in Axios).

servers:
  - url: http://localhost:5000

tags:
  - name: Auth
    description: Registration and login
  - name: Admin
    description: Parking lot and user management
  - name: User
    description: User booking flow, vehicles, and exports

paths:
  /api/register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        201:
          description: User registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        400:
          description: Email exists or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/login:
    post:
      tags: [Auth]
      summary: Authenticate user and start session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        200:
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        401:
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/admin/dashboard:
    get:
      tags: [Admin]
      summary: Get all lots with spot breakdown
      responses:
        200:
          description: List of lots and their spots
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  lots:
                    type: array
                    items:
                      $ref: '#/components/schemas/LotWithSpots'
        401:
          $ref: '#/components/responses/Unauthorized'

  /api/admin/lot/add:
    post:
      tags: [Admin]
      summary: Create a new parking lot (and spots)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LotUpsertRequest'
      responses:
        200:
          description: Lot created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  message: { type: string }
                  lot_id: { type: integer }
        400:
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/admin/lot/{lot_id}:
    get:
      tags: [Admin]
      summary: Fetch a parking lot by id
      parameters:
        - $ref: '#/components/parameters/LotIdPath'
      responses:
        200:
          description: Lot data
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  lot:
                    $ref: '#/components/schemas/Lot'
        404:
          description: Lot not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/admin/lot/update:
    post:
      tags: [Admin]
      summary: Update parking lot details
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/LotUpsertRequest'
                - type: object
                  properties:
                    lot_id: { type: integer }
                  required: [lot_id]
      responses:
        200:
          description: Lot updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        400:
          description: Missing fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/admin/lot/delete/{lot_id}:
    post:
      tags: [Admin]
      summary: Delete a lot and its spots
      parameters:
        - $ref: '#/components/parameters/LotIdPath'
      responses:
        200:
          description: Lot deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        404:
          description: Lot not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/admin/spot/{lot_id}/{spot_id}/{status}:
    get:
      tags: [Admin]
      summary: Return spot metadata (used for dialogs)
      parameters:
        - $ref: '#/components/parameters/LotIdPath'
        - name: spot_id
          in: path
          required: true
          schema: { type: integer }
        - name: status
          in: path
          required: true
          schema: { type: string }
      responses:
        200:
          description: Spot info
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  spot:
                    type: object
                    properties:
                      spot_id: { type: integer }
                      lot_id: { type: integer }
                      status: { type: string }
        400:
          description: Missing parameters

  /api/admin/spot/delete:
    post:
      tags: [Admin]
      summary: Remove a parking spot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                spot_id: { type: integer }
                lot_id: { type: integer }
              required: [spot_id, lot_id]
      responses:
        200:
          description: Spot deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        400:
          description: Missing identifiers

  /api/admin/users:
    get:
      tags: [Admin]
      summary: List registered users
      responses:
        200:
          description: User list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserSummary'

  /api/admin/summary:
    get:
      tags: [Admin]
      summary: Aggregated lot and spot statistics
      responses:
        200:
          description: Metrics for dashboard charts
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  total_lots: { type: integer }
                  total_spots: { type: integer }
                  status_summary:
                    type: object
                    additionalProperties:
                      type: integer
                  lot_spot_data:
                    type: array
                    items:
                      type: array
                      items:
                        oneOf:
                          - type: integer
                          - type: string

  /api/user/dashboard:
    get:
      tags: [User]
      summary: Fetch user parking history and availability snapshot
      responses:
        200:
          description: Dashboard datasets
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  user_parking_data:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserParkingRecord'
                  availability_data:
                    type: array
                    items:
                      $ref: '#/components/schemas/LotAvailability'
        401:
          $ref: '#/components/responses/Unauthorized'

  /api/user/bookSpot:
    post:
      tags: [User]
      summary: Fetch next spot and vehicle list for booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                lot_id: { type: integer }
                locationName: { type: string }
              required: [lot_id, locationName]
      responses:
        200:
          description: Booking context info
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  lot_id: { type: integer }
                  spot_id: { type: integer }
                  user_id: { type: integer }
                  location_name: { type: string }
                  vehicles:
                    type: array
                    items: { type: string }
        401:
          $ref: '#/components/responses/Unauthorized'

  /api/user/addVehicle:
    get:
      tags: [User]
      summary: Return current user id (utility call)
      responses:
        200:
          description: Logged-in user data
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  user_id: { type: integer }
    post:
      tags: [User]
      summary: Register a vehicle for the logged-in user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                vehicle_number: { type: string }
              required: [vehicle_number]
      responses:
        200:
          description: Vehicle created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        400:
          description: Missing field or duplicate vehicle
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/user/confirmBooking:
    post:
      tags: [User]
      summary: Confirm a reservation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                locationName: { type: string }
                lot_id: { type: integer }
                spot_id: { type: integer }
                vehicle_number: { type: string }
              required: [locationName, lot_id, spot_id, vehicle_number]
      responses:
        200:
          description: Reservation persisted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        400:
          description: Validation error

  /api/user/releaseSpot:
    post:
      tags: [User]
      summary: Calculate release cost preview
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                spot_id: { type: integer }
                vehicle_number: { type: string }
                lot_id: { type: integer }
                parking_time: { type: string, format: date-time }
              required: [spot_id, vehicle_number, lot_id, parking_time]
      responses:
        200:
          description: Cost breakdown
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  spot_id: { type: integer }
                  vehicle_number: { type: string }
                  lot_id: { type: integer }
                  parking_time: { type: string }
                  release_time: { type: string }
                  total_cost: { type: number }
        400:
          description: Missing fields

  /api/user/confirmRelease:
    post:
      tags: [User]
      summary: Finalize spot release
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                spot_id: { type: integer }
                lot_id: { type: integer }
                vehicle_number: { type: string }
              required: [spot_id, lot_id, vehicle_number]
      responses:
        200:
          description: Release completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        400:
          description: Missing fields

  /api/user/exportCSV:
    post:
      tags: [User]
      summary: Trigger async parking history export
      responses:
        200:
          description: Export enqueued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        401:
          $ref: '#/components/responses/Unauthorized'

components:
  parameters:
    LotIdPath:
      name: lot_id
      in: path
      required: true
      schema:
        type: integer

  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    SuccessResponse:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }

    ErrorResponse:
      type: object
      properties:
        success: { type: boolean, default: false }
        message: { type: string }

    RegisterRequest:
      type: object
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password }
        fullname: { type: string }
        address: { type: string }
        pincode: { type: string }
        phone_no: { type: string }
      required: [email, password, fullname, address, pincode, phone_no]

    LoginRequest:
      type: object
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password }
      required: [email, password]

    LoginResponse:
      type: object
      properties:
        success: { type: boolean }
        user_id: { type: integer }
        email: { type: string }
        name: { type: string }
        role: { type: string }

    LotUpsertRequest:
      type: object
      properties:
        locationName: { type: string }
        address: { type: string }
        pincode: { type: string }
        pricePerHour: { type: integer }
        maxSpots: { type: integer }
      required: [locationName, address, pincode, pricePerHour, maxSpots]

    Lot:
      type: object
      properties:
        lot_id: { type: integer }
        location_name: { type: string }
        address: { type: string }
        pincode: { type: string }
        price: { type: integer }
        maxSpots: { type: integer }

    LotWithSpots:
      allOf:
        - $ref: '#/components/schemas/Lot'
        - type: object
          properties:
            total: { type: integer }
            occupied: { type: integer }
            available: { type: integer }
            spots:
              type: array
              items:
                type: object
                properties:
                  spot_id: { type: integer }
                  lot_id: { type: integer }
                  status: { type: string }

    UserSummary:
      type: object
      properties:
        id: { type: integer }
        email: { type: string }
        fullname: { type: string }
        address: { type: string }
        pincode: { type: string }

    UserParkingRecord:
      type: object
      properties:
        reserve_id: { type: integer }
        location: { type: string }
        vehicle_number: { type: string }
        parking_timestamp: { type: string }
        status: { type: string }
        lot_id: { type: integer }

    LotAvailability:
      type: object
      properties:
        lot_id: { type: integer }
        location: { type: string }
        availability: { type: integer }